% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaPopulationPlot.R
\name{pcaPopulationPlot}
\alias{pcaPopulationPlot}
\title{PCA Plot for Set of Attributes in a population}
\usage{
pcaPopulationPlot(attributeSet)
}
\arguments{
\item{attributeSet}{The set of attributes}
}
\value{
Returns a PCA plot of different species relatedness
}
\description{
This function will display a PCA plot of n attributes to see how
closely related certain species are related to each other
}
\examples{
pcaPopulationPlot(mtcars)



}
