% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provestiDistDendo.R
\name{provestiDistDendo}
\alias{provestiDistDendo}
\title{Provesti Distance Dendogram}
\usage{
provestiDistDendo(dataSet, rand_n = 10)
}
\arguments{
\item{dataSet}{The set of attributes}

\item{rand_n}{number of individuals/population that use wants to see}
}
\value{
Returns Heatmap plot of all loci suspected of not being in HWE with p
}
\description{
This function will display a dendogram of random n individuals/populations
to visualize  a genetic distance with a given data set.
}
\examples{
provestiDistDendo(nancycats, 10)

}
