% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwHeatPlot.R
\name{hwHeatPlot}
\alias{hwHeatPlot}
\title{Hardy Weinberg Heatplot}
\usage{
hwHeatPlot(dataSet, nPermutations = 100, alpha = 0.05)
}
\arguments{
\item{dataSet}{The set of attributes}

\item{nPermutations}{number of permutations}

\item{alpha}{Alpha value of p}
}
\value{
Returns Heatmap plot of all loci suspected of not being in HWE with p
}
\description{
This function will display a heatmap showing significant departures from HWE
where the pink values are where p are less than or equal to alpha.
}
\examples{
\dontrun{
hwHeatPlot(nancycats, 100, 0.05)
}


}
