% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provestiDistDendo.R
\name{provestiDistDendo}
\alias{provestiDistDendo}
\title{Provesti Distance Dendogram}
\usage{
provestiDistDendo(dataSet, randN = 10)
}
\arguments{
\item{dataSet}{The set of attributes}

\item{randN}{number of individuals/population that use wants to see}
}
\description{
This function will display a dendogram of random n individuals/populations
to visualize  a genetic distance with a given data set.
}
\examples{
\dontrun{
provestiDistDendo(nancycats, 10)
}

}
